% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addError.R
\name{addError}
\alias{addError}
\title{Add error to "observations.}
\usage{
addError(x, cv = NULL, ss = NULL, sdv = NULL, type = "NORMAL")
}
\arguments{
\item{x}{- value(s) to add errors to}

\item{cv}{- error cv}

\item{ss}{- sample size for multinomial distributions}

\item{sdv}{- error standard deviation}

\item{type}{- error distribution (NORMAL, LOGNORMAL, MULTINOMIAL)}
}
\value{
number (or, if x is a vector, vector) with error(s) added
}
\description{
Function to add error to "observations".
}

