% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{asclogisticLn50LnD95}
\alias{asclogisticLn50LnD95}
\title{Calculate an ascending logistic function parameterized by ln(z50) and ln(z95-z50)}
\usage{
asclogisticLn50LnD95(z, lnZ50, lnD95, fsz = 0)
}
\arguments{
\item{z}{- vector of sizes at which to compute selectivities}

\item{lnZ50}{- ln-scale size at which selectivity  = 0.5 (logit-scale mean)}

\item{lnD95}{- ln-scale difference beteen z50 and z95}

\item{fsz}{- if fsz>0, fsz=fully-selected size. if fsz<0, function is normalized to max. if fsz=0, no re-scaling is done}
}
\value{
vector with selectivity values at the elements of z
}
\description{
Function to calculate an ascending logistic function parameterized by ln(z50) and ln(z95-z50)
}
\details{

}

