% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEquilibriumSizeComp.TM.R
\name{calcEquilibriumSizeComp.TM}
\alias{calcEquilibriumSizeComp.TM}
\title{Calculate the equilibrium size composition for one sex of a terminally-molting species with
two maturity states and two shell conditions.}
\usage{
calcEquilibriumSizeComp.TM(mc, R_z, S1_msz, P_sz, Th_sz, T_szz, S2_msz)
}
\arguments{
\item{mc}{- model configuration list object}

\item{R_z}{- vector of equilibrium recruits-at-size}

\item{S1_msz}{- 3d array of pr(survival) from start of year to mating/molting by maturity state, shell condition, size class}

\item{P_sz}{- 2d array of the probability by size class of molting for immature crab, by shell condition}

\item{Th_sz}{- 2d array with pr(molt to maturity|size, molt) for immature crab by shell condition}

\item{T_szz}{- 3d array with size transition matrix for growth by immature crab by shell condition}

\item{S2_msz}{- 3d array of pr(survival) from mating/molting to end of year by maturity state, shell condition, size class}
}
\value{
n_msz, a 3d array with equilibrium numbers-at-size
}
\description{
Function to calculate the equilibrium size composition for one sex of a terminally-molting species with
two maturity states and two shell conditions.
}

