% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcFishingMortalities.R
\name{calcFishingMortalities}
\alias{calcFishingMortalities}
\title{Calculate size-specific fishery capture and retention rates.}
\usage{
calcFishingMortalities(mc, sel_cz, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}

\item{sel_cz}{- selectivity functions}

\item{showPlot}{- flag to show plots}
}
\value{
list with the following elements
sel_fyxmsz - size-specific selectivity
ret_fyxmsz - size-specific retention
devs_fy    - ln-scale deviations in annual capture rates
hm_fxy     - handling mortality rate
cpF_fyxms  - fully-selected fishing capture rate
cpF_fyxmsz - size-specific fishing capture rate
tmF_fyxmsz - size-specific fishing mortality rate
rmF_fyxmsz - size-specific retention mortality rate
dmF_fyxmsz - size-specific discard mortality rate
tmF_yxmsz  - total fishing mortality across all fisheries
}
\description{
Function to calculate size-specific fishery capture and retention rates.
}
\details{

}

