% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcInitSizeComps.R
\name{calcInitSizeComps}
\alias{calcInitSizeComps}
\title{Calculate initial size compositions.}
\usage{
calcInitSizeComps(mc, mp, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration list object}

\item{mp}{- model processes list object}
}
\value{
n_xmsz: 4-d array of initial population abundance by sex/maturity/shell condition/size
}
\description{
Function to calculate initial size compositions.
}
\details{

}

