% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcNaturalMortality.R
\name{calcNaturalMortality}
\alias{calcNaturalMortality}
\title{Calculate natural mortality.}
\usage{
calcNaturalMortality(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}
}
\value{
list  with array M_yxmsz and M_cxm
}
\description{

}

