% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNatZ.Fisheries.R
\name{calcNatZ.Fisheries}
\alias{calcNatZ.Fisheries}
\title{Calculate fishery catches, mortality at size through the model time interval}
\usage{
calcNatZ.Fisheries(mc, mp, N_yxmsz, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration list object}

\item{mp}{- model processes list object}

\item{N_yxmsz}{- numbers at size array}

\item{showPlot}{- flag to show plot.}
}
\value{
list with the following elements:
cpN_fyxms:  5-d array of fishery catches (NOT MORTALITY) by fishery/year/sex/maturity/shell condition
cpB_fyxms:  5-d array of fishery catches (NOT MORTALITY) by fishery/year/sex/maturity/shell condition
cpN_fyxmsz: 6-d array of fishery catches (NOT MORTALITY) by fishery/year/sex/maturity/shell condition/size
cpB_fyxmsz: 6-d array of fishery catches (NOT MORTALITY) by fishery/year/sex/maturity/shell condition/size
dsN_fyxms:  5-d array of fishery discard numbers (NOT MORTALITY) by fishery/year/sex/maturity/shell condition
dsB_fyxms:  5-d array of fishery discard biomass (NOT MORTALITY) by fishery/year/sex/maturity/shell condition
dsN_fyxmsz: 6-d array of fishery discard numbers (NOT MORTALITY) by fishery/year/sex/maturity/shell condition/size
dsB_fyxmsz: 6-d array of fishery discard numbers (NOT MORTALITY) by fishery/year/sex/maturity/shell condition/size
tmN_fyxms:  5-d array of total fishery mortality as numbers by fishery/year/sex/maturity/shell condition
tmB_fyxms:  5-d array of total fishery mortality as biomass by fishery/year/sex/maturity/shell condition
tmN_fyxmsz: 6-d array of total fishery mortality as numbers by fishery/year/sex/maturity/shell condition/size
tmB_fyxmsz: 6-d array of total fishery mortality as numbers by fishery/year/sex/maturity/shell condition/size
rmN_fyxms:  5-d array of retention mortality as numbers by fishery/year/sex/maturity/shell condition
rmB_fyxms:  5-d array of retention mortality as biomass by fishery/year/sex/maturity/shell condition
rmN_fyxmsz: 6-d array of retention mortality by year/sex/maturity/shell condition/size
rmB_fyxmsz: 6-d array of retention mortality by year/sex/maturity/shell condition/size
dmN_fyxms:  5-d array of discard mortality as numbers by fishery/year/sex/maturity/shell condition
dmB_fyxms:  5-d array of discard mortality as biomass by fishery/year/sex/maturity/shell condition
dmN_fyxmsz: 6-d array of discard mortality by fishery/year/sex/maturity/shell condition/size
dmB_fyxmsz: 6-d array of discard mortality by fishery/year/sex/maturity/shell condition/size
}
\description{
Function to calculate fishery catches, mortality at size through the model time interval
}
\details{
None.
}

