% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNatZ.R
\name{calcNatZ}
\alias{calcNatZ}
\title{Project population numbers at size forward through the model time interval}
\usage{
calcNatZ(mc, mp, iN_xmsz, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration list object}

\item{mp}{- model processes list object}

\item{iN_xmsz}{- initial numbers at size array}
}
\value{
list with two elements:
MB_yx   - 2d array with mature biomass at mating time by year/sex
N_yxms  - 4-d array of population numbers by year/sex/maturity/shell condition
B_yxms  - 4-d array of population biomass by year/sex/maturity/shell condition
N_yxmsz - 5-d array of population numbers by year/sex/maturity/shell condition/size
}
\description{
Function to project population numbers at size forward through the model time interval
}
\details{
None.
}

