% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNatZ.Surveys.R
\name{calcNatZ.Surveys}
\alias{calcNatZ.Surveys}
\title{Calculate survey numbers at size through the model time interval}
\usage{
calcNatZ.Surveys(mc, mp, N_yxmsz, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration list object}

\item{mp}{- model processes list object}

\item{N_yxmsz}{- initial numbers at size array}
}
\value{
list with the following elements:
N_vyxms: 5-d array of survey numbers by year/sex/maturity/shell condition
B_vyxms: 5-d array of survey biomass by year/sex/maturity/shell condition
N_vyxmsz: 6-d array of survey numbers by year/sex/maturity/shell condition/size
B_vyxmsz: 6-d array of survey biomass by year/sex/maturity/shell condition/size
}
\description{
Function to calculate survey numbers at size through the model time interval
}
\details{
None.
}

