% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcPrMolt.R
\name{calcPrMolt}
\alias{calcPrMolt}
\title{Calculate probability at size of molting for immature crab by sex, shell condition.}
\usage{
calcPrMolt(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}

\item{showPlot}{- flag to show plots}
}
\value{
prMolt_yxsz
}
\description{

}

