% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcPrMoltToMaturity.R
\name{calcPrMoltToMaturity}
\alias{calcPrMoltToMaturity}
\title{Calculate probability-at-size (pre-molt) for immature crab that a molt is to maturity, by sex and shell condition.}
\usage{
calcPrMoltToMaturity(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}

\item{showPlot}{- flag to show plots}
}
\value{
prMoltToMat_yxsz
}
\description{

}

