% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcRecruitment.R
\name{calcRatZ.All}
\alias{calcRatZ.All}
\title{Calculate proportions recruiting-at-size over all years}
\usage{
calcRatZ.All(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}

\item{showPlot}{- flag (T/F) to show plots}
}
\value{
R_yz: d array with annual proportions of crab recruiting by size
}
\description{

}

