% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRecruitment.R
\name{calcRatZ}
\alias{calcRatZ}
\title{Calculate proportions recruiting-at-size for one set of parameters}
\usage{
calcRatZ(mc, tb, showPlot = FALSE)
}
\arguments{
\item{mc}{- model configuration list object}

\item{tb}{- time block or inits list object}

\item{showPlot}{- flag (T/F) to show plots}
}
\value{
1d array with proportions of crab recruiting by size
}

