% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcRecruitment.R
\name{calcRecruitment}
\alias{calcRecruitment}
\title{Calculate recruitment-at-size (in millions) by year, sex}
\usage{
calcRecruitment(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}

\item{showPlot}{- flag (T/F) to show plots}
}
\value{
R_yxz: 3d array with numbers of crab (in millions) recruiting by year/sex/size
}
\description{

}

