% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStateTransitionMatrices.R
\name{calcStateTransitionMatrices}
\alias{calcStateTransitionMatrices}
\title{Calculate the state transition matrices for a single sex.}
\usage{
calcStateTransitionMatrices(mc, S1_msz, P_sz, Th_sz, T_szz, S2_msz)
}
\arguments{
\item{mc}{- model configuration list object}

\item{S1_msz}{- 3d array of pr(survival) from start of year to mating/molting by maturity state, shell condition, size class}

\item{P_sz}{- 2d array of the probability by size class of molting for immature crab, by shell condition}

\item{Th_sz}{- 2d array with pr(molt to maturity|size, molt) for immature crab by shell condition}

\item{T_szz}{- 3d array with size transition matrix for growth by immature crab by shell condition}

\item{S2_msz}{- 3d array of pr(survival) from mating/molting to end of year by maturity state, shell condition, size class}
}
\value{
list with state transition matrices as named elements:
 A : imm, new -> imm, new
 B : imm, old -> imm, new
 C : imm, new -> imm, old
 D : imm, old -> imm, old
 E : imm, new -> mat, new
 F : imm, old -> mat, new
 G : mat, new -> mat, old
 H : mat, old -> mat, old
}
\description{
Function to calculate the state transition matrices for a single sex.
}

