% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSurveyCatchabilities.R
\name{calcSurveyCatchabilities}
\alias{calcSurveyCatchabilities}
\title{Calculate size-specific survey catchabilities.}
\usage{
calcSurveyCatchabilities(mc, sel_cz, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}

\item{sel_cz}{- selectivity functions}

\item{showPlot}{- flag to show plots}
}
\value{
list with the following elements:\cr
sel_vyxmsz - size-specific selectivity\cr
devs_vy  - ln-scale annual deviations in Q\cr
Q_vxy    - fully-selected survey catchability\cr
Q_vyxms  - fully-selected survey catchability\cr
Q_vyxmsz - size-specific survey catchability
}
\description{
Function to calculate size-specific survey catchabilities.
}

