% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcWatZ.R
\name{calcWatZ}
\alias{calcWatZ}
\title{Calculate weight-at-size by year, sex, maturity state and shell condition.}
\usage{
calcWatZ(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}
}
\value{
list with elements
W_cxmz: 4d array with weight-at-size in KG by time block/sex/maturity state (in grams)
W_yxmsz: 5d array with weight-at-size in KG by year/sex/maturity state/shell condition (in kg)
}
\description{

}
\details{
Input parameters are for weight-at-size in grams, but converted to kg so biomass
is in 1000s t.
}

