% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcZTM.R
\name{calcZTM}
\alias{calcZTM}
\title{Calculate size transition matrices by year, sex, shell condition}
\usage{
calcZTM(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}
}
\value{
list with elements:
T_yxszz: 5d array with size transition matrices by year/sex/shell condition
}
\description{
Function to calculate size transition matrices by year, sex, shell condition
}
\details{

}

