% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compareSizeCompsGG.R
\name{compareSizeCompsGG}
\alias{compareSizeCompsGG}
\title{Compare multiple size comps.}
\usage{
compareSizeCompsGG(n_xmsz = NULL, title = "",
  ylab = "Abundance (millions)", showPlot = TRUE)
}
\arguments{
\item{n_xmsz}{- array (or list of arrays) dimensioned xmsz}

\item{title}{- title for plot}

\item{showPlot}{- flag to show plot immediately}
}
\value{
ggplot2 object
}
\description{
Function to compare multiple size comps.
}
\details{

}

