% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{dbllogistic5095}
\alias{dbllogistic5095}
\title{Calculate a double logistic function parameterized by z50 and D95=z95-z50 for ascending/descending limbs}
\usage{
dbllogistic5095(z, ascZ50, ascD95, dscZ50, dscD95, fsz = 0)
}
\arguments{
\item{z}{- vector of sizes at which to compute selectivities}

\item{ascZ50}{- ascending logistic size at which selectivity  = 0.5 (logit-scale mean)}

\item{ascD95}{- increment from z50 to z95 on ascending limb}

\item{dscZ50}{- descending logistic size at which selectivity  = 0.5 (logit-scale mean)}

\item{dscD95}{- increment from z50 to z95 on descending limb}

\item{fsz}{- if fsz>0, fsz=fully-selected size. if fsz<0, will use max(z) as fsz. if fsz=0, no re-scaling is done}
}
\value{
vector with selectivity values at the elements of z
}
\description{
Function to calculate a double logistic function parameterized by z50 and D95=z95-z50 for ascending/descending limbs
}
\details{

}

