% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{dbllogisticLn50Ln95}
\alias{dbllogisticLn50Ln95}
\title{Calculate a double logistic function parameterized by ln-scale versions of z50 and D95=z95-z50 for ascending/descending limbs}
\usage{
dbllogisticLn50Ln95(z, ascLnZ50, ascLnD95, dscLnZ50, dscLnD95, fsz = 0)
}
\arguments{
\item{z}{- vector of sizes at which to compute selectivities}

\item{ascLnZ50}{- ln-scale ascending logistic size at which selectivity  = 0.5 (logit-scale mean)}

\item{ascLnD95}{- ln-scale increment from z50 to z95 on ascending limb}

\item{dscLnZ50}{- ln-scale descending logistic size at which selectivity  = 0.5 (logit-scale mean)}

\item{dscLnD95}{- ln-scale increment from z50 to z95 on descending limb}

\item{fsz}{- if fsz>0, fsz=fully-selected size. if fsz<0, will use max(z) as fsz. if fsz=0, no re-scaling is done}
}
\value{
vector with selectivity values at the elements of z
}
\description{
Function to calculate a double logistic function parameterized by ln-scale z50 and D95=z95-z50 for ascending/descending limbs
}
\details{

}

