% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{dbllogisticLnD50}
\alias{dbllogisticLnD50}
\title{Calculate a double logistic function parameterized by z50 and slope for ascending/descending limbs}
\usage{
dbllogisticLnD50(z, ascZ50, ascSlope, lnD50, dscSlope, fsz = 0)
}
\arguments{
\item{z}{- vector of sizes at which to compute selectivities}

\item{ascZ50}{- ascending logistic size at which selectivity  = 0.5 (logit-scale mean)}

\item{ascSlope}{- ascending logistic slope at z50}

\item{lnD50}{- ln-scale difference between ascending, descending limb z50s}

\item{dscSlope}{- descending logistic slope at z50}

\item{fsz}{- if fsz>0, fsz=fully-selected size. if fsz<0, function is normalized to max. if fsz=0, no re-scaling is done}
}
\value{
vector with selectivity values at the elements of z
}
\description{
Function to calculate a double logistic function parameterized by z50 and slope for ascending/descending limbs
}
\details{

}

