% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dimArray.R
\name{dimArray}
\alias{dimArray}
\title{Get an array}
\usage{
dimArray(mc, str, val = 0)
}
\arguments{
\item{mc}{- model configurqtion object}

\item{str}{- string describing array indices}
}
\value{
array of appropriate dimensions filled with default value(s)
}
\description{

}
\details{
To match arrays in TCSAM2015, indices reflecting parameter combinations
(such as for selectivity functions) should be indicated as "pc_xxx",
where xxx is the rsim dimension name (e.g., 'selfcns'). The resulting array will use
'pc' as the corresponding dimension name
}

