% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.CanonicalFormat.R
\name{getMDFR.CanonicalFormat}
\alias{getMDFR.CanonicalFormat}
\title{Convert dataframe to canonical format}
\usage{
getMDFR.CanonicalFormat(mdfr)
}
\arguments{
\item{mdfr}{- dataframe to convert}
}
\value{
dataframe in canonical format
}
\description{
Function to convert dataframe to canonical format.
}
\details{
returns a dataframe in canonical format
}

