% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Fisheries.SelFcns.R
\name{getMDFR.Fisheries.SelFcns}
\alias{getMDFR.Fisheries.SelFcns}
\title{Get fishery selectivity functions from model results from rsimTCSAM model runs as a dataframe}
\usage{
getMDFR.Fisheries.SelFcns(rsims, verbose = FALSE, cast = "x")
}
\arguments{
\item{rsims}{- single rsimTCSAM object, rsimTCSAM.resLst object, or named list of the latter}

\item{verbose}{- flag (T/F) to print debug info}

\item{cast}{- casting formula for excluding x,m,s factor levels from an average-at-size across unspecified factors}
}
\value{
dataframe in canonical format
}
\description{
Function to get fishery selectivity functions from model results from TCSAM02 model runs as a dataframe.
}
\details{
Extracts fishery selectivity functions.
}

