% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Pop.Abundance.R
\name{getMDFR.Pop.Abundance}
\alias{getMDFR.Pop.Abundance}
\title{Get population abundance time series from model results from rsimTCSAM model runs as a dataframe}
\usage{
getMDFR.Pop.Abundance(rsims, cast = "y+x", verbose = FALSE)
}
\arguments{
\item{rsims}{- single rsimTCSAM.rep object, rsimTCSAM.resLst object, or named list of the latter}

\item{cast}{- casting formula for excluding y,x,m,s,z factor levels from an average-at-size across unspecified factors}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe in canonical format
}
\description{
Function to get population abundance time series from model results from rsimTCSAM model runs as a dataframe.
}
\details{
Extracts the estimated population abundance time series.
}

