% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Pop.GrowthTransitionMatrices.R
\name{getMDFR.Pop.GrowthTransitionMatrices}
\alias{getMDFR.Pop.GrowthTransitionMatrices}
\title{Get growth transition matrices from model results from rsimTCSAM model runs as a dataframe}
\usage{
getMDFR.Pop.GrowthTransitionMatrices(rsims, verbose = FALSE)
}
\arguments{
\item{rsims}{- single rsimTCSAM results object, or named list of such}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe in canconical format.
Note that 'z' is pre-molt size, 'zp' is post-molt size.
}
\description{
Function to get growth transition matrices from model results from rsimTCSAM model runs as a dataframe.
}
\details{
Extracts growth transition matrices.
}

