% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Pop.MeanGrowthIncrements.R
\name{getMDFR.Pop.MeanGrowthIncrements}
\alias{getMDFR.Pop.MeanGrowthIncrements}
\title{Get mean growth increments from model results from TCSAM2015 and rsimTCSAM model runs as a dataframe}
\usage{
getMDFR.Pop.MeanGrowthIncrements(rsims, verbose = FALSE)
}
\arguments{
\item{rsims}{- single rsimTCSAM results object, or named list of such}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe in canonical format.
}
\description{
Function to get mean growth increments from model results from TCSAM2015 and rsimTCSAM model runs as a dataframe.
}
\details{
Extracts mean growth increments.
}

