% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Pop.NaturalMortality.R
\name{getMDFR.Pop.NaturalMortality}
\alias{getMDFR.Pop.NaturalMortality}
\title{Get natural mortality rates from model results from rsimTCSAM model runs as a dataframe}
\usage{
getMDFR.Pop.NaturalMortality(rsims, type = c("M_cxm", "M_yxm", "M_yxmsz"),
  verbose = FALSE)
}
\arguments{
\item{rsims}{- single rsimTCSAM results object, or named list of such}

\item{type}{- flag indicating which M's to extract ('M_cxm', 'M_yxm' or 'M_yxmsz')}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe in canonical format
}
\description{
Function to get natural mortality rates from model results from rsimTCSAM model runs as a dataframe.
}
\details{
Extracts natural mortality rates
}

