% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Pop.PrM2M.R
\name{getMDFR.Pop.PrM2M}
\alias{getMDFR.Pop.PrM2M}
\title{Get molt-to-maturity ogives from model results from rsimTCSAM model runs as a dataframe}
\usage{
getMDFR.Pop.PrM2M(rsims, verbose = FALSE)
}
\arguments{
\item{rsims}{- single rsimTCSAM results object, or named list of such}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe in canonical format.
}
\description{
Function to get molt-to-maturity ogives from model results from rsimTCSAM model runs as a dataframe.
}
\details{
Extracts molt-to-maturity ogives.
}

