% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Pop.RecSizeDistribution.R
\name{getMDFR.Pop.RecSizeDistribution}
\alias{getMDFR.Pop.RecSizeDistribution}
\title{Get recruitment size distribution from model results from rsimTCSAM model runs as a dataframe}
\usage{
getMDFR.Pop.RecSizeDistribution(rsims, verbose = FALSE)
}
\arguments{
\item{rsims}{- single rsimTCSAM results object, or named list of such}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe in canonical format.
}
\description{
Function to get recruitment size distribution from model results from rsimTCSAM model runs as a dataframe.
}
\details{
Extracts mean growth increments.
}

