% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Pop.SexRatio.R
\name{getMDFR.Pop.SexRatio}
\alias{getMDFR.Pop.SexRatio}
\title{Get recruitment sex ratio from model results from rsimTCSAM model runs as a dataframe}
\usage{
getMDFR.Pop.SexRatio(rsims, verbose = FALSE)
}
\arguments{
\item{rsims}{- single rsimTCSAM results object, or named list of such}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe in canonical format.
}
\description{
Function to get recruitment sex ratio from model results from rsimTCSAM model runs as a dataframe.
}
\details{
Extracts recruitment sex ratio.
}

