% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.R
\name{getMDFR}
\alias{getMDFR}
\title{Get model arrays as a melted dataframe from rsimTCSAM models}
\usage{
getMDFR(path, rsims, verbose = FALSE)
}
\arguments{
\item{path}{- path in models to requested array (using '/' as separator for list levels)}

\item{rsims}{- single rsimTCSAM results object, or named list of such}

\item{verbose}{- flag (T/F) to print diagnostics}
}
\value{
Melted dataframe (ala package reshape2).
}
\description{
Function to get model objects as a melted dataframe from rsimTCSAM models.
}
\details{
Returned dataframe is a melted (ala reshape2) version of the requested array in
canonical format. The array value is in column 'val'. Uses \code{reshape2::melt(...)}.
}

