% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parseMC.Growth.R
\name{parseMC.Growth}
\alias{parseMC.Growth}
\title{Parse growth section of a model configuration file.}
\usage{
parseMC.Growth(rsp, i, dims)
}
\arguments{
\item{rsp}{- parsed text list from model configuration file}

\item{i}{- index to start of growth section in rsp}

\item{dims}{- model configuration dims list}
}
\value{
list with
i - index starting next section
params - list object with growth time blocks
}
\description{
Function to parse the 'growth' section of a model configuration file.
}
\details{

}

