% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMC.NaturalMortality.R
\name{parseMC.NaturalMortality}
\alias{parseMC.NaturalMortality}
\title{Parse natural mortality section of a model configuration file.}
\usage{
parseMC.NaturalMortality(rsp, i, dims)
}
\arguments{
\item{rsp}{- parsed text list from model configuration file}

\item{i}{- index to start of natural mortality section in rsp}

\item{dims}{- model configuration dims list}
}
\value{
list with
i - index starting next section
params - list object with natural mortality time blocks
}
\description{
Function to parse the 'natural mortality' section of a model configuration file.
}

