% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parseMC.OutputInfo.R
\name{parseMC.OILine}
\alias{parseMC.OILine}
\title{Parse output info for a fishery/survey data type from a model configuration file.}
\usage{
parseMC.OILine(rsp, i)
}
\arguments{
\item{rsp}{- parsed text list from model configuration file}

\item{i}{- index to start of output info section in rsp}
}
\value{
list with elements
flag - flag (TRUE/FALSE) to output this data type
aggType - TCSAM aggregation type (e.g., BY_TOTAL, BY_SEX)
errType - TCSAM error structure type (e.g., LOGNORMAL, MULTINOMIAL)
err - output cv or sample size
wgt - likelihood weight
}
\description{
Function to read output info for a fishery/survey data type from a model configuration file.
}
\details{


Note that i is not incremented here.
}

