% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMC.PrM2M.R
\name{parseMC.PrM2M}
\alias{parseMC.PrM2M}
\title{Parse molt-to-maturity section of a model configuration file.}
\usage{
parseMC.PrM2M(rsp, i, dims)
}
\arguments{
\item{rsp}{- parsed text list from model configuration file}

\item{i}{- index to start of molt-to-maturity section in rsp}

\item{dims}{- model configuration dims list}
}
\value{
list with
i - index starting next section
params - list object with molt-to-maturity time blocks
}
\description{
Function to parse the 'molt-to-maturity' section of a model configuration file.
}

