% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parseMC.Recruitment.R
\name{parseMC.Recruitment}
\alias{parseMC.Recruitment}
\title{Parse recruitment section of a model configuration file.}
\usage{
parseMC.Recruitment(rsp, i, dims)
}
\arguments{
\item{rsp}{- parsed text list from model configuration file}

\item{i}{- index to start of recruitment section in rsp}

\item{dims}{- model configuration dims list}
}
\value{
list with
i - index starting next section
params - list object with recruitment time blocks
}
\description{
Function to parse the 'recruitment' section of a model configuration file.
}
\details{

}

