% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parseMC.SelFcns.R
\name{parseMC.SelFcns}
\alias{parseMC.SelFcns}
\title{Parse selectivity section of a model configuration file.}
\usage{
parseMC.SelFcns(rsp, i, dims)
}
\arguments{
\item{rsp}{- parsed text list from model configuration file}

\item{i}{- index to start of selectivity section in rsp}

\item{dims}{- model configuration dims list}
}
\value{
list with
i - index starting next section
selfcns - list object with selectivity info (type,params,label)
}
\description{
Function to parse the 'selectivity' section of a model configuration file.
}
\details{

}

