% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parseMC.WatZ.R
\name{parseMC.WatZ}
\alias{parseMC.WatZ}
\title{Parse weight-at-size section of a model configuration file.}
\usage{
parseMC.WatZ(rsp, i, dims)
}
\arguments{
\item{rsp}{- parsed text list from model configuration file}

\item{i}{- index to start of weight-at-size section in rsp}

\item{dims}{- model configuration dims list}
}
\value{
list with
i - index starting next section
params - list object with weight-at-size time blocks
}
\description{
Function to parse the 'weight-at-size' section of a model configuration file.
}
\details{

}

