% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{plogis}
\alias{plogis}
\title{Calculate the logistic function}
\usage{
plogis(z, z50, sd, fsz)
}
\arguments{
\item{z}{- vector of sizes at which to compute selectivities}

\item{z50}{- size at which selectivity  = 0.5 (logit-scale mean)}

\item{sd}{- standard deviation in selectivity (logit-scale standard deviation)}

\item{fsz}{- if fsz>0, fsz=fully-selected size. if fsz<0, will use max(z) as fsz. if fsz=0, no re-scaling is done}
}
\value{
vector with selectivity values at the elements of z
}
\description{
Function to calculate the logistic function
}
\details{

}

