% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readModelConfiguration.TCSAM.R
\name{readModelConfiguration.TCSAM}
\alias{readModelConfiguration.TCSAM}
\title{Select and read a model configuration file for a TCSAM model.}
\usage{
readModelConfiguration.TCSAM(fn = NULL, ext = "dat")
}
\arguments{
\item{fn}{- the file to read (NULL brings up a file chooser dialog)}

\item{ext}{- extension of type of file to select (if fn is NULL)}
}
\value{
model configuration list object
}
\description{
Function to select and read a model configuration file for a TCSAM.
}

