% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeImmOS.R
\name{removeImmOS}
\alias{removeImmOS}
\title{Remove rows marked as immature, old shell from a dataframe}
\usage{
removeImmOS(mdfr, maturity = "m", shell_condition = "s")
}
\arguments{
\item{mdfr}{- datarame from which to remove the immature, old shell factor combination}

\item{maturity}{- name of column specifying maturity state}

\item{shell_condition}{- name of column specifying shell condition}
}
\value{
dataframe with immature, old shell factor combination removed
}
\description{
Function to remove rows marked as immature, old shell from a dataframe.
}

