% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcNatZ.R
\name{runOneYear.TM}
\alias{runOneYear.TM}
\title{Project one sex of the population one time step forward.}
\usage{
runOneYear.TM(mc, R_z, S1_msz, P_sz, Th_sz, T_szz, S2_msz, n_msz)
}
\arguments{
\item{mc}{- model configuration list object}

\item{R_z}{- vector of recruits-at-size}

\item{S1_msz}{- 3d array of pr(survival) from start of year to mating/molting by maturity state, shell condition, size class}

\item{P_sz}{- 2d array of the probability by size class of molting for immature crab, by shell condition}

\item{Th_sz}{- 2d array with pr(molt to maturity|size, molt) for immature crab by shell condition}

\item{T_szz}{- 3d array with size transition matrix for growth by immature crab by shell condition}

\item{S2_msz}{- 3d array of pr(survival) from mating/molting to end of year by maturity state, shell condition, size class}

\item{n_msz}{- 3d array with initial nubers at size}
}
\value{
np_msz, a 3d array with projected numbers-at-size
}
\description{
Function to project one sex of the population one time step forward.
}
\details{

}

