% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/runSim.TCSAM.R
\name{runSim.TCSAM}
\alias{runSim.TCSAM}
\title{Run the TCSAM simulation model to produce a TCSAM2015 input file.}
\usage{
runSim.TCSAM(fn = "tcsam.input.dat", showPlot = TRUE)
}
\arguments{
\item{fn}{- filename for output file (input file to gmacs)}

\item{showPlot}{- flag (T/F) to show plots}
}
\value{
list with elements:
mc - model configuration list object
mp - model processes list object
mr - model results list object
}
\description{
Function to run the TCSAM simulation model to produce a TCASM2015 input file.
}
\details{

}

