% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/writeSim.TCSAM.R
\name{writeSim.TCSAM}
\alias{writeSim.TCSAM}
\title{Write simulation results to TCSAM2015 input file.}
\usage{
writeSim.TCSAM(mc, mp, mr, out.dir = ".", showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration list object}

\item{mp}{- model processes list object}

\item{mr}{- model results list object}

\item{out.dir}{- folder for output files}

\item{showPlot}{- flag to show plots}
}
\value{
list with elements F_list and S_list
}
\description{
Function to write simulation results to TCSAM2015 input file.
}
\details{

}

