% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/writeSim.TCSAM.Surveys.R
\name{writeSim.TCSAM.Surveys}
\alias{writeSim.TCSAM.Surveys}
\title{Write model survey results to output file.}
\usage{
writeSim.TCSAM.Surveys(mc, mp, mr, conn, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration list object}

\item{mp}{- model processes list object}

\item{mr}{- model results list object}

\item{conn}{- open file connection for output}

\item{showPlot}{- flag to show plots}
}
\description{
Function to write model survey results to output file.
}
\details{

}

