% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSim.TCSAM02.R
\name{writeSim.TCSAM02}
\alias{writeSim.TCSAM02}
\title{Write simulation results to TCSAM02 input file2.}
\usage{
writeSim.TCSAM02(mc, mp, mr, out.dir = ".", showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration list object}

\item{mp}{- model processes list object}

\item{mr}{- model results list object}

\item{out.dir}{- folder for output files}

\item{showPlot}{- flag to show plots}
}
\value{
list with elements F_list and S_list
}
\description{
Function to write simulation results to TCSAM02 input files.
}

