% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSim.TCSAM02.Surveys.R
\name{writeSim.TCSAM02.Surveys}
\alias{writeSim.TCSAM02.Surveys}
\title{Write model survey results to TCSAM02 input file.}
\usage{
writeSim.TCSAM02.Surveys(mc, mp, mr, fnSrvs, out.dir = ".", showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration list object}

\item{mp}{- model processes list object}

\item{mr}{- model results list object}

\item{fnSrvs}{- files to write survey data to}

\item{out.dir}{- folder to write survey data to}

\item{showPlot}{- flag to show plots}
}
\description{
Function to write model survey results to TCSAM02 input file.
}

