% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSim.TCSAM2013.R
\name{writeSim.TCSAM2013}
\alias{writeSim.TCSAM2013}
\title{Write simulation results to TCSAM2013 input file.}
\usage{
writeSim.TCSAM2013(mc, mp, mr, out.dir = ".", showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration list object}

\item{mp}{- model processes list object}

\item{mr}{- model results list object}

\item{out.dir}{- folder for output files}

\item{showPlot}{- flag to show plots}
}
\value{
list with elements F_list and S_list
}
\description{
Function to write simulation results to TCSAM2013 input file.
}

