% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSim.TCSAM2017.Fisheries.R
\name{writeSim.TCSAM2017.Fisheries}
\alias{writeSim.TCSAM2017.Fisheries}
\title{Write rsim fisheries results to TCSAM2017 input file.}
\usage{
writeSim.TCSAM2017.Fisheries(mc, mp, mr, fnFshs, out.dir = ".",
  showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration list object}

\item{mp}{- model processes list object}

\item{mr}{- model results list object}

\item{fnFshs}{- files to write fishery data to}

\item{out.dir}{- folder to write survey data to}

\item{showPlot}{- flag to show plots}
}
\value{
list with retained, discarded, captured numbers and biomass by f,y,x,s
}
\description{
Function to write model fisheries results to TCSAM2017 input file.
}

