% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSim.TCSAM2017.Surveys.R
\name{writeSim.TCSAM2017.Surveys}
\alias{writeSim.TCSAM2017.Surveys}
\title{Write model survey results to TCSAM2017 input file.}
\usage{
writeSim.TCSAM2017.Surveys(mc, mp, mr, fnSrvs, out.dir = ".",
  showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration list object}

\item{mp}{- model processes list object}

\item{mr}{- model results list object}

\item{fnSrvs}{- files to write survey data to}

\item{out.dir}{- folder to write survey data to}

\item{showPlot}{- flag to show plots}
}
\description{
Function to write model survey results to TCSAM2017 input file.
}

