% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRecruitment.R
\name{calcRatZ.All}
\alias{calcRatZ.All}
\title{Calculate proportions recruiting-at-size over all time blocks}
\usage{
calcRatZ.All(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}

\item{showPlot}{- flag (T/F) to show plots}
}
\value{
R_cz: d array with annual proportions (by time block) of crab recruiting by size
}

