% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Fisheries.CatchBiomass.R
\name{getMDFR.Fisheries.CatchBiomass}
\alias{getMDFR.Fisheries.CatchBiomass}
\title{Get fishery catch biomass time series from model results from rsimTCSAM model runs as a dataframe}
\usage{
getMDFR.Fisheries.CatchBiomass(rsims, category = c("captured", "discarded",
  "retained", "discard mortality", "total mortality", "index"), cast = "y+x",
  verbose = FALSE)
}
\arguments{
\item{rsims}{- single rsimTCSAM.rep object, rsimTCSAM.resLst object, or named list of the latter}

\item{category}{- 'captured','discarded','retained', 'discard mortality', 'total mortality', or 'index'}

\item{cast}{- casting formula for excluding y,x,m,s,z factor levels from an average-at-size across unspecified factors}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe in canonical format
}
\description{
Function to get fishery catch biomass time series from model results from rsimTCSAM model runs as a dataframe.
}
\details{
Extracts the estimated fishery biomass time series.
}

