% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResLst.R
\name{getResLst}
\alias{getResLst}
\title{Create a rsimTCSAM.resLst object from a model run}
\usage{
getResLst(path = NULL)
}
\arguments{
\item{path}{- path to .RData object containing model run}
}
\value{
a rsimTCSAM.resLst object.
}
\description{
Function to create a rsimTCSAM.resLst object from a model run.
}
\details{
Uses \code{tcltk::tk_chose.file} to open a file dialog to select model .RData file
if path is NULL. A rsimTCSAM.resLst object is a list with elements
\itemize{
 \item{rep - a rsimTCSAM object, or NULL}
}
}

