% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runSim.TCSAM.R
\name{runSim.TCSAM}
\alias{runSim.TCSAM}
\title{Run the TCSAM simulation model to produce a TCSAM2015 input file.}
\usage{
runSim.TCSAM(out.dir = ".", seed = NULL, fnd = "devs.csv",
  showPlot = TRUE, pdf = NULL, width = 8, height = 6)
}
\arguments{
\item{out.dir}{- directory path for output files (input files to TCSAM2015)}

\item{seed}{- seed for random number generator (NULL generates system seed)}

\item{fnd}{- filename for devs output}

\item{showPlot}{- flag (T/F) to show plots}

\item{pdf}{- if not NULL, a filename to output plots to in pdf format}

\item{width}{- width of pdf output (inches)}

\item{height}{- height of pdf output (inches)}
}
\value{
list with elements:
seed - RNG seed used
mc - model configuration list object (see readModelConfiguration.TCSAM(...))
mp - model processes list object    (see calcModelProcesses(...))
mr - model results list object      (see runModel(...))
}
\description{
Function to run the TCSAM simulation model to produce a TCASM2015 input file.
}
\details{

}

