% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/writeSim.TCSAM.Fisheries.R
\name{writeSim.TCSAM.Fisheries}
\alias{writeSim.TCSAM.Fisheries}
\title{Write model fisheries results to output file.}
\usage{
writeSim.TCSAM.Fisheries(mc, mp, mr, fnFshs, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration list object}

\item{mp}{- model processes list object}

\item{mr}{- model results list object}

\item{fnFshs}{- files to write fishery data to}

\item{showPlot}{- flag to show plots}
}
\value{
list with retained, discarded, captured numbers and biomass by f,y,x
}
\description{
Function to write model fisheries results to output file.
}
\details{

}

